backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
google-genai
gitpython
python-dotenv
cors
